Got rid of all unnecessary generics and moved to a safe marker type
All the network logic is now inside of the processor class and aren't leaking channels to other classes anymore.
Redirect logic is implemented.
And engines don't have access to markers, sockets, routers and etc. anymore
Engines are safe to be used as sandboxes for running code sent from client and not mess any important logic.